bioinformatics: Remove strelka.
A newer version is available in Guix proper.  This version fails to
build, requires a much older version of boost, would probably not work
with the much more recent htslib and samtools, and it depends on
Python 2.  All of these are good reasons to remove these packages.

* guix-science-nonfree/packages/bioinformatics.scm (strelka-1.0.14,
strelka-1.0.15): Remove variables.