machine-learning: python-jaxlib/wheel-with-cuda11: Fix CUDA build.
When --config=cuda_plugin is passed to Bazel it evaluates the
build:cuda_plugin lines in .bazelrc.  These *disable* CUDA support.
Since the GPU plugin is experimental and not expected to work anyway,
there is really no point in building it.

* guix-science-nonfree/packages/machine-learning.scm
(python-jaxlib/wheel-with-cuda11)[arguments]: Remove 'build-gpu-plugin
phase; remove configure flags from run-command, bazel-arguments, and
.jax_configure.bazelrc.
(python-jax-cuda-plugin-with-cuda11): Remove variable.
(python-jax-with-cuda11)[propagated-inputs]: Remove
python-jax-cuda-plugin-with-cuda11.